(split)bug: none
ByteUtil.copy should return a long, not an int.  Also, try to fix a bug that would copy more data than requested when a limit is specified that is not a multiple of 8192.

Copied from Perforce
 Change: 148720

+++ src/java/com/zimbra/cs/imap/ImapPartSpecifier.java
@@ -135,7 +135,7 @@ class ImapPartSpecifier {
                 ps.print(binary ? "~{" : "{");  ps.print(length);  ps.write('}');
                 if (os != null) {
                     os.write(ImapHandler.LINE_SEPARATOR_BYTES);
-                    int written = ByteUtil.copy(is, false, os, false);
+                    long written = ByteUtil.copy(is, false, os, false);
                     assert(written == length);
                 }
             }
